build(make): add test-integration + test-all Makefile targets
- make test-integration: runs 38 integration tests (real DuckDB/DemoAdapter)
- make test-all: runs unit + integration together
- make test: unit tests only (unchanged)
- make coverage: unit tests with 94% coverage
- 382 unit + 38 integration = 420 total tests all green

+++ src/hexawyn/infrastructure/memory/sql/search_similar.sql
@@ -11,7 +11,7 @@ SELECT
     solution,
     severity,
     weight,
-    array_cosine_similarity(embedding, ?::DOUBLE[?]) * weight
+    array_cosine_similarity(embedding, ?::FLOAT[?]) * weight
         / ln(age_days + 2) AS score
 FROM incidents
 WHERE cluster_name = ?
